MainFrame and AppConfig improvements
Removed the idea of "Base directory" for AppConfig (still need to remove
code about this anti-feature)
Improve the use of Pluma by MainFrame

Warning: a crash has been found in official Pluma framework. I have
fixed it, but while the author do not fix it, ask me for fixed code.